+++ guess-game-server/src/main/java/guess/domain/statistics/olap/measure/MvpsQuantityMeasure.java
@@ -13,7 +13,7 @@ public class MvpsQuantityMeasure extends Measure<Speaker> {
     }
 
     @Override
-    public Long calculateValue() {
-        return (long) entities.size();
+    public long calculateValue() {
+        return entities.size();
     }
 }
